c: configure: no kernel feat checks at build time
Today, most people doesn't compile .c applications. Build servers
generate the binaries, making some dependencies on other userspace
programs but no dependencies on the kernel versions.

There is no need because the kernel ensure the uAPI is stable. If
something is not available, the feature is not supported but if it is
there, the way we use it should not change.

Checking MPTCP support should then be done at running time.

We should then remove the build time check to avoid misuses.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>