add 2 Rust examples (socket2 & hyper)
Nowadays Rust is more and more used to build applications
and tools. So it can be a good idea to have examples
showing how to enable MPTCP on some of the most used
libraries/frameworks.

For now, both examples rely on the git version of socket2
because they just merged a PR to support MPTCP. So we
need to wait until they create a new release before using
the version tag in Cargo.toml.

-- Socket2

This is the most basic example. Socket2 a low-level
library providing a safe API to create socket in Rust.
Thus this example is pretty much the same as the C one.

-- Hyper server

This one is more complex, Hyper is a low-level HTTP
library that allows its users to create web server or
HTTP client. It is one of the most used HTTP library and
is then a perfect candidate for an example.

In the Hyper example, I've not introduced a helper
function to create the socket (as with the socket2
example) because, in theory, we only need to create one
socket, thus the helper function would only be called once
(so not really useful, I guess).

Signed-off-by: Martin Andre <martin.andre@tessares.net>